requirements.txt
torch
absl-py==0.6.1
arff==0.9
astor==0.7.1
backcall==0.1.0
bleach==3.0.2
boto3==1.9.62
botocore==1.12.62
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
configparser==3.7.1
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
defusedxml==0.5.0
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
filterpy==1.4.5
gast==0.2.0
grpcio==1.16.1
h5py==2.9.0
hjson==3.0.1
idna==2.7
imageio==2.4.1
imagesize==1.1.0
imgaug==0.2.7
imutils==0.5.1
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.2
munch==2.3.2
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numba==0.42.0
numpy==1.15.4
opencv-contrib-python==3.4.3.18
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pika==0.12.0
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.6.1
ptyprocess==0.6.0
py-cpuinfo==4.0.0
Pygments==2.2.0
pyparsing==2.2.2
pypng==0.0.19
python-dateutil==2.7.4
pytorch-ignite==0.1.2
pytz==2018.6
PyWavelets==1.0.1
pyzmq==17.1.2
qtconsole==4.4.2
rawpy==0.13.1
requests==2.20.0
s3transfer==0.1.13
sacred==0.7.4
scikit-image==0.14.1
scikit-learn==0.20.0
scikit-multilearn==0.1.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
#Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.12.0
snakeviz==1.0.0
tensorboard==1.12.1
tensorboardX==1.6
#tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
#torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
#xgboost==0.81
